allow cpt tables to updated with new data.
The update code does a "biased" average of the old and new cpt tables.
Bias is a factor between 0.0 and 1.0 and controls how much the new cpt values
skew to the old cpt values.  0.0 means always use the new values, and 1.0
means only use old cpt values.  bias=.5 is a standard average.